+++ evalml/pipelines/classification/xgboost_multiclass.py
@@ -2,7 +2,7 @@ import numpy as np
 from skopt.space import Integer, Real
 
 from evalml.model_types import ModelTypes
-from evalml.pipelines import PipelineBase
+from evalml.pipelines import MulticlassClassificationPipeline
 from evalml.pipelines.components import (
     OneHotEncoder,
     RFClassifierSelectFromModel,
@@ -12,11 +12,11 @@ from evalml.pipelines.components import (
 from evalml.problem_types import ProblemTypes
 
 
-class XGBoostPipeline(PipelineBase):
-    """XGBoost Pipeline for both binary and multiclass classification"""
+class XGBoostMulticlassPipeline(MulticlassClassificationPipeline):
+    """XGBoost Pipeline for multiclass classification"""
     name = "XGBoost Classifier w/ One Hot Encoder + Simple Imputer + RF Classifier Select From Model"
     model_type = ModelTypes.XGBOOST
-    problem_types = [ProblemTypes.BINARY, ProblemTypes.MULTICLASS]
+    problem_types = [ProblemTypes.MULTICLASS]
 
     hyperparameters = {
         "eta": Real(0, 1),
@@ -27,7 +27,7 @@ class XGBoostPipeline(PipelineBase):
         "percent_features": Real(.01, 1),
     }
 
-    def __init__(self, objective, eta, min_child_weight, max_depth, impute_strategy,
+    def __init__(self, eta, min_child_weight, max_depth, impute_strategy,
                  percent_features, number_features, n_estimators=10, n_jobs=-1, random_state=0):
 
         imputer = SimpleImputer(impute_strategy=impute_strategy)
@@ -45,7 +45,6 @@ class XGBoostPipeline(PipelineBase):
                                       min_child_weight=min_child_weight,
                                       n_estimators=n_estimators)
 
-        super().__init__(objective=objective,
-                         component_list=[enc, imputer, feature_selection, estimator],
+        super().__init__(component_list=[enc, imputer, feature_selection, estimator],
                          n_jobs=n_jobs,
                          random_state=random_state)