+++ evalml/pipelines/regression/catboost.py
@@ -1,12 +1,12 @@
 from skopt.space import Integer, Real
 
 from evalml.model_types import ModelTypes
-from evalml.pipelines import PipelineBase
+from evalml.pipelines import RegressionPipeline
 from evalml.pipelines.components import CatBoostRegressor, SimpleImputer
 from evalml.problem_types import ProblemTypes
 
 
-class CatBoostRegressionPipeline(PipelineBase):
+class CatBoostRegressionPipeline(RegressionPipeline):
     """
     CatBoost Pipeline for regression problems.
     CatBoost is an open-source library and natively supports categorical features.
@@ -23,7 +23,7 @@ class CatBoostRegressionPipeline(PipelineBase):
         "max_depth": Integer(1, 8),
     }
 
-    def __init__(self, objective, impute_strategy, n_estimators, eta,
+    def __init__(self, impute_strategy, n_estimators, eta,
                  max_depth, number_features, bootstrap_type=None,
                  n_jobs=-1, random_state=0):
         # note: impute_strategy must support both string and numeric data
@@ -33,7 +33,6 @@ class CatBoostRegressionPipeline(PipelineBase):
                                       max_depth=max_depth,
                                       bootstrap_type=bootstrap_type,
                                       random_state=random_state)
-        super().__init__(objective=objective,
-                         component_list=[imputer, estimator],
+        super().__init__(component_list=[imputer, estimator],
                          n_jobs=n_jobs,
                          random_state=random_state)