+++ evalml/pipelines/regression/random_forest.py
@@ -2,7 +2,7 @@ import numpy as np
 from skopt.space import Integer, Real
 
 from evalml.model_types import ModelTypes
-from evalml.pipelines import PipelineBase
+from evalml.pipelines import RegressionPipeline
 from evalml.pipelines.components import (
     OneHotEncoder,
     RandomForestRegressor,
@@ -12,7 +12,7 @@ from evalml.pipelines.components import (
 from evalml.problem_types import ProblemTypes
 
 
-class RFRegressionPipeline(PipelineBase):
+class RFRegressionPipeline(RegressionPipeline):
     """Random Forest Pipeline for regression problems"""
     name = "Random Forest Regressor w/ One Hot Encoder + Simple Imputer + RF Regressor Select From Model"
     model_type = ModelTypes.RANDOM_FOREST
@@ -25,7 +25,7 @@ class RFRegressionPipeline(PipelineBase):
         "percent_features": Real(.01, 1)
     }
 
-    def __init__(self, objective, n_estimators, max_depth, impute_strategy, percent_features,
+    def __init__(self, n_estimators, max_depth, impute_strategy, percent_features,
                  number_features, n_jobs=-1, random_state=0):
 
         imputer = SimpleImputer(impute_strategy=impute_strategy)
@@ -42,7 +42,6 @@ class RFRegressionPipeline(PipelineBase):
                                           max_depth=max_depth,
                                           n_jobs=n_jobs)
 
-        super().__init__(objective=objective,
-                         component_list=[enc, imputer, feature_selection, estimator],
+        super().__init__(component_list=[enc, imputer, feature_selection, estimator],
                          n_jobs=n_jobs,
                          random_state=random_state)