+++ evalml/pipelines/regression_pipeline.py
@@ -1,0 +1,5 @@
+from evalml.pipelines import PipelineBase
+
+
+class RegressionPipeline(PipelineBase):
+    pass