+++ evalml/tests/automl_tests/test_auto_classification_search.py
@@ -97,7 +97,7 @@ def test_binary_auto(X_y):
     X, y = X_y
     automl = AutoClassificationSearch(objective="recall", multiclass=False, max_pipelines=5)
     automl.search(X, y, raise_errors=True)
-    y_pred = automl.best_pipeline.predict(X)
+    y_pred = automl.best_pipeline.predict(X, "recall")
 
     assert len(np.unique(y_pred)) == 2
 
@@ -127,6 +127,7 @@ def test_multi_auto(X_y_multi):
     expected_additional_objectives = get_objectives('multiclass')
     objective_in_additional_objectives = next((obj for obj in expected_additional_objectives if obj.name == objective.name), None)
     expected_additional_objectives.remove(objective_in_additional_objectives)
+
     for expected, additional in zip(expected_additional_objectives, automl.additional_objectives):
         assert type(additional) is type(expected)
 
@@ -233,7 +234,7 @@ def test_describe_pipeline_objective_ordered(X_y, capsys):
     out_stripped = " ".join(out.split())
 
     objectives = [get_objective(obj) for obj in automl.additional_objectives]
-    objectives_names = [automl.objective.name] + [obj.name for obj in objectives if obj.name not in ["ROC", "Confusion Matrix"]]
+    objectives_names = [obj.name for obj in objectives if obj.name not in ["ROC", "Confusion Matrix"]]
     expected_objective_order = " ".join(objectives_names)
 
     assert err == ''