+++ evalml/tests/pipeline_tests/regression_pipeline_tests/test_catboost_regression.py
@@ -8,8 +8,7 @@ from evalml.pipelines import CatBoostRegressionPipeline
 
 
 def test_catboost_init():
-    objective = R2()
-    clf = CatBoostRegressionPipeline(objective=objective, impute_strategy='mean', n_estimators=1000, number_features=0,
+    clf = CatBoostRegressionPipeline(impute_strategy='mean', n_estimators=1000, number_features=0,
                                      bootstrap_type='Bayesian', eta=0.03, max_depth=6, random_state=2)
     expected_parameters = {'impute_strategy': 'mean', 'eta': 0.03, 'n_estimators': 1000, 'max_depth': 6, 'bootstrap_type': 'Bayesian'}
     assert clf.parameters == expected_parameters
@@ -27,22 +26,27 @@ def test_catboost_regression(X_y_reg):
     sk_score = sk_pipeline.score(X, y)
 
     objective = R2()
-    clf = CatBoostRegressionPipeline(objective=objective, n_estimators=1000, eta=0.03, number_features=X.shape[1],
+    clf = CatBoostRegressionPipeline(n_estimators=1000, eta=0.03, number_features=X.shape[1],
                                      bootstrap_type='Bayesian', max_depth=6, impute_strategy='mean', random_state=0)
     clf.fit(X, y)
-    clf_score = clf.score(X, y)
+    clf_score = clf.score(X, y, [objective])
     y_pred = clf.predict(X)
 
     np.testing.assert_almost_equal(y_pred, sk_pipeline.predict(X), decimal=5)
     np.testing.assert_almost_equal(sk_score, clf_score[0], decimal=5)
 
+    # testing objective parameter passed in does not change results
+    clf.fit(X, y, objective)
+    y_pred_with_objective = clf.predict(X, objective)
+    np.testing.assert_almost_equal(y_pred, y_pred_with_objective, decimal=5)
+
 
 def test_cbr_input_feature_names(X_y_categorical_regression):
     X, y = X_y_categorical_regression
     objective = R2()
-    clf = CatBoostRegressionPipeline(objective=objective, impute_strategy='most_frequent', n_estimators=1000,
+    clf = CatBoostRegressionPipeline(impute_strategy='most_frequent', n_estimators=1000,
                                      number_features=len(X.columns), bootstrap_type='Bayesian',
                                      eta=0.03, max_depth=6, random_state=0)
-    clf.fit(X, y)
+    clf.fit(X, y, objective)
     assert len(clf.feature_importances) == len(X.columns)
     assert not clf.feature_importances.isnull().all().all()