+++ evalml/tests/pipeline_tests/test_pipelines.py
@@ -4,7 +4,11 @@ import pytest
 
 from evalml.model_types import ModelTypes
 from evalml.objectives import FraudCost, Precision
-from evalml.pipelines import LogisticRegressionPipeline, PipelineBase
+from evalml.pipelines import (
+    LogisticRegressionBinaryPipeline,
+    LogisticRegressionMulticlassPipeline,
+    PipelineBase
+)
 from evalml.pipelines.components import (
     LogisticRegressionClassifier,
     OneHotEncoder,
@@ -42,10 +46,15 @@ def test_serialization(X_y, tmpdir):
     path = os.path.join(str(tmpdir), 'pipe.pkl')
     objective = Precision()
 
-    pipeline = LogisticRegressionPipeline(objective=objective, penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]))
-    pipeline.fit(X, y)
+    pipeline = LogisticRegressionBinaryPipeline(penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]))
+    pipeline.fit(X, y, objective)
     save_pipeline(pipeline, path)
-    assert pipeline.score(X, y) == load_pipeline(path).score(X, y)
+    assert pipeline.score(X, y, [objective]) == load_pipeline(path).score(X, y, [objective])
+
+    pipeline = LogisticRegressionMulticlassPipeline(penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]))
+    pipeline.fit(X, y, objective)
+    save_pipeline(pipeline, path)
+    assert pipeline.score(X, y, [objective]) == load_pipeline(path).score(X, y, [objective])
 
 
 @pytest.fixture
@@ -54,8 +63,8 @@ def pickled_pipeline_path(X_y, tmpdir):
     path = os.path.join(str(tmpdir), 'pickled_pipe.pkl')
     MockPrecision = type('MockPrecision', (Precision,), {})
     objective = MockPrecision()
-    pipeline = LogisticRegressionPipeline(objective=objective, penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]))
-    pipeline.fit(X, y)
+    pipeline = LogisticRegressionBinaryPipeline(penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]))
+    pipeline.fit(X, y, objective)
     save_pipeline(pipeline, path)
     return path
 
@@ -66,9 +75,9 @@ def test_load_pickled_pipeline_with_custom_objective(X_y, pickled_pipeline_path)
     with pytest.raises(NameError):
         MockPrecision()  # noqa: F821: ignore flake8's "undefined name" error
     objective = Precision()
-    pipeline = LogisticRegressionPipeline(objective=objective, penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]))
-    pipeline.fit(X, y)
-    assert load_pipeline(pickled_pipeline_path).score(X, y) == pipeline.score(X, y)
+    pipeline = LogisticRegressionBinaryPipeline(penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]))
+    pipeline.fit(X, y, 'precision')
+    assert load_pipeline(pickled_pipeline_path).score(X, y, [objective]) == pipeline.score(X, y, [objective])
 
 
 def test_reproducibility(X_y):
@@ -80,19 +89,19 @@ def test_reproducibility(X_y):
         amount_col=10
     )
 
-    clf = LogisticRegressionPipeline(objective=objective, penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
-    clf.fit(X, y)
+    clf = LogisticRegressionBinaryPipeline(penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
+    clf.fit(X, y, objective)
 
-    clf_1 = LogisticRegressionPipeline(objective=objective, penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
-    clf_1.fit(X, y)
+    clf_1 = LogisticRegressionBinaryPipeline(penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
+    clf_1.fit(X, y, objective)
 
-    assert clf_1.score(X, y) == clf.score(X, y)
+    assert clf_1.score(X, y, [objective]) == clf.score(X, y, [objective])
 
 
 def test_indexing(X_y):
     X, y = X_y
-    clf = LogisticRegressionPipeline(objective='recall', penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
-    clf.fit(X, y)
+    clf = LogisticRegressionBinaryPipeline(penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
+    clf.fit(X, y, 'recall')
 
     assert isinstance(clf[0], OneHotEncoder)
     assert isinstance(clf['One Hot Encoder'], OneHotEncoder)
@@ -108,13 +117,13 @@ def test_indexing(X_y):
 
 def test_describe(X_y):
     X, y = X_y
-    lrp = LogisticRegressionPipeline(objective='recall', penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
+    lrp = LogisticRegressionBinaryPipeline(penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
     assert lrp.describe(True) == {'C': 1.0, 'impute_strategy': 'mean', 'penalty': 'l2'}
 
 
 def test_name(X_y):
     X, y = X_y
-    clf = LogisticRegressionPipeline(objective='recall', penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
+    clf = LogisticRegressionBinaryPipeline(penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
     assert clf.name == 'Logistic Regression Classifier w/ One Hot Encoder + Simple Imputer + Standard Scaler'
 
 
@@ -124,7 +133,7 @@ def test_estimator_not_last(X_y):
     class MockLogisticRegressionPipeline(PipelineBase):
         name = "Mock Logistic Regression Pipeline"
 
-        def __init__(self, objective, penalty, C, impute_strategy,
+        def __init__(self, penalty, C, impute_strategy,
                      number_features, n_jobs=-1, random_state=0):
             imputer = SimpleImputer(impute_strategy=impute_strategy)
             enc = OneHotEncoder()
@@ -133,55 +142,63 @@ def test_estimator_not_last(X_y):
                                                      penalty=penalty,
                                                      C=C,
                                                      n_jobs=-1)
-            super().__init__(objective=objective, component_list=[enc, imputer, estimator, scaler], n_jobs=n_jobs, random_state=random_state)
+            super().__init__(component_list=[enc, imputer, estimator, scaler], n_jobs=n_jobs, random_state=random_state)
 
     err_msg = "A pipeline must have an Estimator as the last component in component_list."
     with pytest.raises(ValueError, match=err_msg):
-        MockLogisticRegressionPipeline(objective='recall', penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
+        MockLogisticRegressionPipeline(penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
 
 
 def test_multi_format_creation(X_y):
     X, y = X_y
-    clf = PipelineBase('precision', component_list=['Simple Imputer', 'One Hot Encoder', StandardScaler(), 'Logistic Regression Classifier'], n_jobs=-1, random_state=0)
+    clf = PipelineBase(component_list=['Simple Imputer', 'One Hot Encoder', StandardScaler(), 'Logistic Regression Classifier'], n_jobs=-1, random_state=0)
     correct_components = [SimpleImputer, OneHotEncoder, StandardScaler, LogisticRegressionClassifier]
     for component, correct_components in zip(clf.component_list, correct_components):
         assert isinstance(component, correct_components)
     assert clf.model_type == ModelTypes.LINEAR_MODEL
     assert clf.problem_types == [ProblemTypes.BINARY, ProblemTypes.MULTICLASS]
 
-    clf.fit(X, y)
-    clf.score(X, y)
+    clf.fit(X, y, 'precision')
+    clf.score(X, y, ['precision'])
     assert not clf.feature_importances.isnull().all().all()
 
 
 def test_multiple_feature_selectors(X_y):
     X, y = X_y
-    clf = PipelineBase('precision', component_list=['Simple Imputer', 'One Hot Encoder', 'RF Classifier Select From Model', StandardScaler(), 'RF Classifier Select From Model', 'Logistic Regression Classifier'], n_jobs=-1, random_state=0)
+    clf = PipelineBase(component_list=['Simple Imputer', 'One Hot Encoder', 'RF Classifier Select From Model', StandardScaler(), 'RF Classifier Select From Model', 'Logistic Regression Classifier'], n_jobs=-1, random_state=0)
     correct_components = [SimpleImputer, OneHotEncoder, RFClassifierSelectFromModel, StandardScaler, RFClassifierSelectFromModel, LogisticRegressionClassifier]
     for component, correct_components in zip(clf.component_list, correct_components):
         assert isinstance(component, correct_components)
     assert clf.model_type == ModelTypes.LINEAR_MODEL
     assert clf.problem_types == [ProblemTypes.BINARY, ProblemTypes.MULTICLASS]
 
-    clf.fit(X, y)
-    clf.score(X, y)
+    clf.fit(X, y, 'precision')
+    clf.score(X, y, ['precision'])
     assert not clf.feature_importances.isnull().all().all()
+
+
+def test_score_with_empty_list_of_objectives(X_y):
+    X, y = X_y
+    clf = LogisticRegressionBinaryPipeline(penalty='l2', C=1.0, impute_strategy='mean', number_features=len(X[0]), random_state=0)
+    clf.fit(X, y)
+    with pytest.raises(IndexError):
+        clf.score(X, y, [])
 
 
 def test_n_jobs(X_y):
     with pytest.raises(ValueError, match='n_jobs must be an non-zero integer*.'):
-        PipelineBase('precision', component_list=['Simple Imputer', 'One Hot Encoder', StandardScaler(), 'Logistic Regression Classifier'],
+        PipelineBase(component_list=['Simple Imputer', 'One Hot Encoder', StandardScaler(), 'Logistic Regression Classifier'],
                      n_jobs='5', random_state=0)
 
     with pytest.raises(ValueError, match='n_jobs must be an non-zero integer*.'):
-        PipelineBase('precision', component_list=['Simple Imputer', 'One Hot Encoder', StandardScaler(), 'Logistic Regression Classifier'],
+        PipelineBase(component_list=['Simple Imputer', 'One Hot Encoder', StandardScaler(), 'Logistic Regression Classifier'],
                      n_jobs=0, random_state=0)
 
-    assert PipelineBase('precision', component_list=['Simple Imputer', 'One Hot Encoder', StandardScaler(), 'Logistic Regression Classifier'],
+    assert PipelineBase(component_list=['Simple Imputer', 'One Hot Encoder', StandardScaler(), 'Logistic Regression Classifier'],
                         n_jobs=-4, random_state=0)
 
-    assert PipelineBase('precision', component_list=['Simple Imputer', 'One Hot Encoder', StandardScaler(), 'Logistic Regression Classifier'],
+    assert PipelineBase(component_list=['Simple Imputer', 'One Hot Encoder', StandardScaler(), 'Logistic Regression Classifier'],
                         n_jobs=4, random_state=0)
 
-    assert PipelineBase('precision', component_list=['Simple Imputer', 'One Hot Encoder', StandardScaler(), 'Logistic Regression Classifier'],
+    assert PipelineBase(component_list=['Simple Imputer', 'One Hot Encoder', StandardScaler(), 'Logistic Regression Classifier'],
                         n_jobs=None, random_state=0)